+++ src/frontend/packages/lib_video/src/components/common/VideoWidgetProvider/widgets/SharedLiveMedia/SharedLiveMediaItem/DeleteSharedLiveMediaButton/index.tsx
@@ -32,7 +32,7 @@ export const DeleteSharedLiveMediaButton = ({
       icon={
         <BinSVG
           height="18px"
-          iconColor={colorsTokens['info-500']}
+          iconColor={colorsTokens['danger-200']}
           width="14px"
         />
       }